made install.sh a little better and starting to work on a animation

+++ install.sh
@@ -1,14 +1,37 @@
-#!/bin/sh 
+#!/bin/bash
 
-set -xe
+set -e
 
 # RUN THIS FILE ONLY IF YOU WANT TO ADD EXECUTABLE PRIVELAGES TO ANY FILE USING A NEVER CHANGING DIRECTORY
 
-sudo rm -rf /usr/bin/hustle
-sudo mkdir /usr/bin/hustle
-sudo cp hustle.py /usr/bin/hustle
-sudo cp stdlib.hsle /usr/bin/hustle 
-sudo cp std -r /usr/bin/hustle
-sudo cp arrow_strings -r /usr/bin/hustle
-sudo cp keywords -r /usr/bin/hustle
-sudo cp ops -r /usr/bin/hustle
+if [ "$EUID" -ne 0 ]
+then
+        printf "Please run this program as root by typing \"sudo ./install.sh\" \n"
+        exit 1
+else
+       printf "[CMD] sudo rm -rf /usr/bin/hustle \n"
+       sudo rm -rf /usr/bin/hustle
+       
+       printf "[CMD] sudo mkdir /usr/bin/hustle \n"
+       sudo mkdir /usr/bin/hustle
+       
+       printf "[CMD] sudo cp hustle.py /usr/bin/hustle \n"
+       sudo cp hustle.py /usr/bin/hustle
+       
+       printf "[CMD] sudo cp stdlib.hsle /usr/bin/hustle \n" 
+       sudo cp stdlib.hsle /usr/bin/hustle
+       
+       printf "[CMD] sudo cp std -r /usr/bin/hustle \n"
+       sudo cp std -r /usr/bin/hustle
+       
+       printf "[CMD] sudo cp arrow_strings -r /usr/bin/hustle \n"
+       sudo cp arrow_strings -r /usr/bin/hustle
+
+       printf "[CMD] sudo cp keywords -r /usr/bin/hustle \n"
+       sudo cp keywords -r /usr/bin/hustle
+
+       printf "[CMD] sudo cp ops -r /usr/bin/hustle \n"
+       sudo cp ops -r /usr/bin/hustle
+
+       printf "\nInstallation Complete, You can now run hustle files as executables"
+fi
